search-provider: replace signals
This fixes a problem we were seeing, where not all data was being
preloaded in y-s-p. We should be instead waiting for
YELP_DOCUMENT_SIGNAL_CONTENTS signal.

Semantics of INFO and CONTENTS signal are not documented, so
it is not clear whether we can get all the page ids from
document after getting the INFO signal.Because of this we
rather wait for the CONTENTS signal.