Desktop file: temporary hack for downstream name and icon
Hardcoded the downstream name and comment (including translations),
and icon, to work around build infrastructure problem that is
not properly merging the content in eos-shell-content.

We ran out of time to fix properly for the eos-3.2.0 release,
so we will temporarily live with the following limitations:
- Future translations in Transifex will be ignored
- Name and subtitle will be in English (rather than any upstream
  translation) for any languages that we do not fully support
- Missing description and screenshots in app center

This commit should be reverted once we fix the integration
with the eoscontent debhelper.

https://phabricator.endlessm.com/T17879