search-provider: Serialize an instance of GFileIcon, not a GThemedIcon
Yelp's search provider will normally look up for its own icons using
its own instance of the default GtkIconTheme, which it would have
extended to include the non-standard /usr/share/yelp/icons directory
in the search path, as that's where it installs its own icons.

That works well inside Yelp because it can find those icons by name,
but it does not work in the shell's side because a serialized themed
icon will just contain the name, which won't be enough for the shell
to find the actual icon for as it will live in a non-standard path.

Let's easily solve this by serializing a GFileIcon instead, which will
remove the need to lookup for the right icon based on its name in the
shell's side, as a GFileIcon will already know the path to the actual
file, already retrieved in the provider's side.

Note that this approach forces us to specify a size in the provider,
but that's probably fine since a .svg will normally be retrieved, and
because we're asking for 32px icons, which is what the shell demmands.

https://phabricator.endlessm.com/T18065